# swaggerDocs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CUJAE LDAP API",
        "version": "1.0.0",
        "description": "API para la administración de usuarios LDAP de la CUJAE",
        "contact": {
            "name": "Ahmed González",
            "email": "ahmediglez@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:4005"
        }
    ],
...
    "paths": {
        "/api/v1/searchByDN": {
            "get": {
                "tags": [
                    "DN"
                ],
                "summary": "Search by Distinguished Name (DN).",
                "description": "Retrieve information by providing a Distinguished Name (DN).",
                "operationId": "searchByDN",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "baseDN",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The Distinguished Name (DN) to search for."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved information.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LDAPResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request. Base DN is missing."
                    },
                    "401": {
                        "description": "Unauthorized. Requires admin role."
                    },
                    "500": {
                        "description": "Internal Server Error. Failed to fetch information."
                    }
                }
            }
        },
        "/api/v1/groups/byName/{group}": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get a LDAP group by its name.",
                "description": "Retrieve information about an LDAP group by specifying its name in the URL path. This endpoint requires authentication to access group information.",
                "operationId": "getGroupByName",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The name of the LDAP group to retrieve."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "LDAP group information retrieved successfully."
                    },
                    "401": {
                        "description": "Unauthorized. The user is not authenticated."
                    },
                    "500": {
                        "description": "An error occurred while fetching group information."
                    }
                }
            }
        },
        "/api/v1/groups/getChilds": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get child groups of a specified base DN.",
                "description": "Retrieve child groups of a specified base DN. This endpoint is restricted to administrators and requires authentication.",
                "operationId": "getChildGroups",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "baseDN",
                        "required": true,
                        "description": "The base DN (Distinguished Name) for which to retrieve child groups.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "baseDN": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Child groups retrieved successfully."
                    },
                    "401": {
                        "description": "Unauthorized. The user is not authenticated."
                    },
                    "403": {
                        "description": "Forbidden. The user does not have sufficient privileges to access this endpoint."
                    },
                    "500": {
                        "description": "An error occurred while fetching child groups."
                    }
                }
            }
        },
        "/api/v1/groups/byType/{type}": {
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get a group by type.",
                "description": "Retrieve a group by its type using the specified DN. This endpoint requires authentication.",
                "operationId": "getGroupByType",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "type",
                        "required": true,
                        "description": "The type of the group to retrieve.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "body",
                        "name": "dn",
                        "description": "The DN (Distinguished Name) to search for the group.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "dn": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Group retrieved successfully by type."
                    },
                    "401": {
                        "description": "Unauthorized. The user is not authenticated."
                    },
                    "500": {
                        "description": "An error occurred while fetching the group by type."
                    }
                }
            }
        },
...
        "/api/v1/group/{group}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a list of users in a specific group.",
                "description": "Retrieve a list of users in the specified group based on query parameters.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The group name for which to retrieve users."
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        },
                        "description": "The page number."
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        },
                        "description": "The number of items per page."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of users in the group."
                    },
                    "500": {
                        "description": "An error occurred."
                    }
                }
            }
        },
...
    }
}